services/comprehension/bing-spelling-api/requirements.txt
# Function dependencies, for example:
# package>=version
flask>=1.1.1
pyspellchecker>=0.5.2
pytest>=5.2.2
requests>=2.22.0
